toutes les fonctionnalités fonctionnent sauf les likes

+++ backend/controllers/sauce.js
@@ -1,38 +1,82 @@
 const Sauce = require("../models/sauce");
+const fs = require("fs");
+/* exports.createSauce = (req, res, next) => {
+  // Permet de creer une nouvelle sauce
+  delete req.body.userId;
+  delete req.body._id;
+  console.log(req.body.sauce);
+  const sauce = new Sauce({
+    ...req.body,
+  });
+  console.log(sauce);
+  sauce
+    .save() // .save() permet d'enregistrer notre objet
+    .then(() => res.status(201).json({ message: "objet enregistré" }))
+    .catch((error) => res.status(400).json({ error }));
+}; */
 
 exports.createSauce = (req, res, next) => {
-  console.log(req.body);
+  console.log(req.body.sauce);
   const sauceObject = JSON.parse(req.body.sauce);
   delete sauceObject._id;
   delete sauceObject._userId;
   const sauce = new Sauce({
     ...sauceObject,
     userId: req.auth.userId,
-    imageUrl: `${req.protocol}://${req.get("host")}/images/${
-      req.file.filename
-    }`,
+    imageUrl: `${req.protocol}://${req.get("host")}/image/${req.file.filename}`,
   });
   sauce
     .save()
     .then(() => {
-      res.status(201).json({ message: "Objet enregistré" });
+      res.status(201).json({ message: "Objet enregistré !" });
     })
     .catch((error) => {
       res.status(400).json({ error });
     });
 };
 exports.modifySauce = (req, res, next) => {
-  // permet de modifier une sauce
-  Sauce.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id }) // .updateOne permet de modifier un élément selectionné avec son id
-    .then(() => res.status(200).json({ message: "object modifié" }))
+  const sauceObject = req.file
+    ? {
+        ...JSON.parse(req.body.sauce),
+        imageUrl: `${req.protocol}://${req.get("host")}/image/${
+          req.file.filename
+        }`,
+      }
+    : { ...req.body };
+  delete sauceObject._userId;
+  Sauce.findOne({ _id: req.params.id })
+    .then((sauce) => {
+      if (sauce.userId != req.auth.userId) {
+        res.status(400).json({ message: "Non-autorisé" });
+      } else {
+        Sauce.updateOne(
+          { _id: req.params.id },
+          { ...sauceObject, _id: req.params.id }
+        )
+          .then(() => res.status(200).json({ message: "objet modifié" }))
+          .catch((error) => ({ error }));
+      }
+    })
     .catch((error) => res.status(400).json({ error }));
 };
 
 exports.deleteSauce = (req, res, next) => {
-  // permet de supprimer une sauce
-  Sauce.deleteOne({ _id: req.params.id }) // .deleteOne supprime un élément qu'on cible avec l'id
-    .then(res.status(200).json({ message: "Sauce supprimé" }))
-    .catch((error) => res.status(400).json({ error }));
+  Sauce.findOne({ _id: req.params.id })
+    .then((sauce) => {
+      if (sauce.userId != req.auth.userId) {
+        res.status(401).json({ message: "Non authorisé " });
+      } else {
+        const filename = sauce.imageUrl.split("/image/")[1];
+        fs.unlink(`image/${filename}`, () => {
+          Sauce.deleteOne({ _id: req.params.id })
+            .then(() => {
+              res.status(200).json({ message: "objet supprimé" });
+            })
+            .catch((error) => res.status(401).json({ error }));
+        });
+      }
+    })
+    .catch((error) => res.status(500).json({ error }));
 };
 
 exports.getOneSauce = (req, res, next) => {
@@ -44,6 +88,7 @@ exports.getOneSauce = (req, res, next) => {
 
 exports.getAllSauce = (req, res, next) => {
   // Permet de voir toutes les sauces
+
   Sauce.find() // .find permet de retrouver tous les éléments
     .then((sauces) => res.status(200).json(sauces))
     .catch((error) => res.status({ error }));
